doc: Remove claim about ISO C
The patch to make -fcommon the default introduces a bogus claim into
the GCC documentation.

-fcommon was claimed to be incompatible with ISO C for preventing
duplicate definitions from being diagnosed.  It does, but as that
elicits undefined behaviour (the requirement that there shall be no
more than one external definition is not a constraint), ISO C does not
require any diagnostic for it.  In the absence of any other rule this
would violate, both -fcommon and -fno-common are fully compatible with
all versions of ISO C.

2019-11-21  Harald van Dijk  <harald@gigawatt.nl>

	* doc/invoke.texi (-fcommon): Remove claim about ISO C.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@278604 138bc75d-0d04-0410-961f-82ee72b054a4